+++ src/ClassicUO.Client/Game/UI/Gumps/ModernOptionsGump.cs
@@ -28,20 +28,11 @@ namespace ClassicUO.Game.UI.Gumps
         private List<SettingsOption> _options = new List<SettingsOption>();
         private Profile profile;
 
-        private string[] GetNamePlatePresetOptions() => new[]
-        {
-            TazLang.Get("nameplate_preset_custom", "Custom"),
-            TazLang.Get("nameplate_preset_orion", "Orion"),
-            TazLang.Get("nameplate_preset_wow_blockybars", "WoW - Blocky Bars"),
-            TazLang.Get("nameplate_preset_wow_cleanhealth", "WoW - Clean Health"),
-            TazLang.Get("nameplate_preset_wow_blockycast", "WoW - Blocky Cast"),
-            TazLang.Get("nameplate_preset_wow_redname", "WoW - Red Name")
-        };
-
         private string[] GetNamePlateBackgroundModeOptions() => new[]
         {
             TazLang.Get("nameplate_background_fixedcolor", "Fixed color"),
-            TazLang.Get("nameplate_background_notorietycolor", "Notoriety color")
+            TazLang.Get("nameplate_background_notorietycolor", "Notoriety color"),
+            TazLang.Get("nameplate_background_entitynotorietycolor", "Entity notoriety color")
         };
 
         private string[] GetNamePlateHealthBarModeOptions() => new[]
@@ -386,7 +377,7 @@ namespace ClassicUO.Game.UI.Gumps
             (
                 new CheckboxWithLabel
                 (TazLang.Get("mog_general_incomingmobiles"), isChecked: profile.ShowNewMobileNameIncoming,
-                    valueChanged: (b) => { profile.ShowNewMobileNameIncoming = b; }), true, page
+                    valueChanged: (b) => { profile.ShowNewMobileNameIncoming = b; SetNamePlatePresetCustom(); }), true, page
             );
 
             content.BlankLine();
@@ -395,7 +386,7 @@ namespace ClassicUO.Game.UI.Gumps
             (
                 new CheckboxWithLabel
                 (TazLang.Get("mog_general_incomingcorpses"), isChecked: profile.ShowNewCorpseNameIncoming,
-                    valueChanged: (b) => { profile.ShowNewCorpseNameIncoming = b; }), true, page
+                    valueChanged: (b) => { profile.ShowNewCorpseNameIncoming = b; SetNamePlatePresetCustom(); }), true, page
             );
 
             content.BlankLine();
@@ -2966,150 +2957,25 @@ namespace ClassicUO.Game.UI.Gumps
 
         private void SetNamePlatePresetCustom()
         {
-            if (profile.NamePlatePreset != NamePlatePreset.Custom)
-            {
-                profile.NamePlatePreset = NamePlatePreset.Custom;
-            }
-
-            NameOverheadGump.InvalidateAllLayouts();
+            NamePlatePresets.SetCustom(profile);
         }
 
-        private void ApplyNamePlatePreset(NamePlatePreset preset)
+        private void ApplyNamePlatePreset(NamePlatePresets.Entry preset)
         {
-            profile.NamePlatePreset = preset;
+            NamePlatePresets.Apply(profile, preset);
+            MainThreadQueue.EnqueueAction(RefreshNamePlateOptions);
+        }
 
-            switch (preset)
-            {
-                case NamePlatePreset.Orion:
-                    profile.NamePlateUseFixedWidth = true;
-                    profile.NamePlateFixedWidth = 160;
-                    profile.NamePlateUseFixedHealthBarWidth = false;
-                    profile.NamePlateHealthBarFixedWidth = 160;
-                    profile.NamePlateHeight = 0;
-                    profile.NamePlateSplitHealthBar = false;
-                    profile.NamePlateCornerRadius = 18;
-                    profile.NamePlateHealthBarMode = NamePlateHealthBarMode.StatusColor;
-                    profile.NamePlateBackgroundMode = NamePlateBackgroundMode.NotorietyColor;
-                    profile.NamePlateBackgroundR = 0;
-                    profile.NamePlateBackgroundG = 0;
-                    profile.NamePlateBackgroundB = 0;
-                    profile.NamePlateHealthBar = true;
-                    profile.NamePlateHealthBarOpacity = 75;
-                    profile.NamePlateOpacity = 70;
-                    profile.NamePlateBorderOpacity = 80;
-                    profile.NamePlateAvoidOverlap = true;
-                    profile.NamePlateHideAtFullHealth = false;
-                    profile.NamePlateHideAtFullHealthInWarmode = false;
-                    profile.NamePlateShowWordOfDeathIcon = false;
-                    profile.NamePlateFont = "avadonian";
-                    profile.NamePlateFontSize = 16;
-                    break;
+        private void RefreshNamePlateOptions()
+        {
+            if (IsDisposed)
+                return;
 
-                case NamePlatePreset.WorldOfWarcraftBlockyBars:
-                    profile.NamePlateUseFixedWidth = true;
-                    profile.NamePlateFixedWidth = 220;
-                    profile.NamePlateUseFixedHealthBarWidth = false;
-                    profile.NamePlateHealthBarFixedWidth = 220;
-                    profile.NamePlateHeight = 44;
-                    profile.NamePlateSplitHealthBar = true;
-                    profile.NamePlateCornerRadius = 2;
-                    profile.NamePlateHealthBarMode = NamePlateHealthBarMode.Green;
-                    profile.NamePlateBackgroundMode = NamePlateBackgroundMode.FixedColor;
-                    profile.NamePlateBackgroundR = 18;
-                    profile.NamePlateBackgroundG = 14;
-                    profile.NamePlateBackgroundB = 14;
-                    profile.NamePlateHealthBar = true;
-                    profile.NamePlateHealthBarOpacity = 100;
-                    profile.NamePlateOpacity = 85;
-                    profile.NamePlateBorderOpacity = 85;
-                    profile.NamePlateAvoidOverlap = true;
-                    profile.NamePlateHideAtFullHealth = false;
-                    profile.NamePlateHideAtFullHealthInWarmode = false;
-                    profile.NamePlateShowWordOfDeathIcon = false;
-                    profile.NamePlateFont = "avadonian";
-                    profile.NamePlateFontSize = 17;
-                    break;
-
-                case NamePlatePreset.WorldOfWarcraftCleanHealth:
-                    profile.NamePlateUseFixedWidth = true;
-                    profile.NamePlateFixedWidth = 220;
-                    profile.NamePlateUseFixedHealthBarWidth = false;
-                    profile.NamePlateHealthBarFixedWidth = 220;
-                    profile.NamePlateHeight = 54;
-                    profile.NamePlateSplitHealthBar = true;
-                    profile.NamePlateCornerRadius = 3;
-                    profile.NamePlateHealthBarMode = NamePlateHealthBarMode.Green;
-                    profile.NamePlateBackgroundMode = NamePlateBackgroundMode.FixedColor;
-                    profile.NamePlateBackgroundR = 20;
-                    profile.NamePlateBackgroundG = 18;
-                    profile.NamePlateBackgroundB = 18;
-                    profile.NamePlateHealthBar = true;
-                    profile.NamePlateHealthBarOpacity = 100;
-                    profile.NamePlateOpacity = 85;
-                    profile.NamePlateBorderOpacity = 90;
-                    profile.NamePlateAvoidOverlap = true;
-                    profile.NamePlateHideAtFullHealth = false;
-                    profile.NamePlateHideAtFullHealthInWarmode = false;
-                    profile.NamePlateShowWordOfDeathIcon = false;
-                    profile.NamePlateFont = "avadonian";
-                    profile.NamePlateFontSize = 18;
-                    break;
-
-                case NamePlatePreset.WorldOfWarcraftBlockyCast:
-                    profile.NamePlateUseFixedWidth = true;
-                    profile.NamePlateFixedWidth = 220;
-                    profile.NamePlateUseFixedHealthBarWidth = false;
-                    profile.NamePlateHealthBarFixedWidth = 220;
-                    profile.NamePlateHeight = 36;
-                    profile.NamePlateSplitHealthBar = true;
-                    profile.NamePlateCornerRadius = 1;
-                    profile.NamePlateHealthBarMode = NamePlateHealthBarMode.Green;
-                    profile.NamePlateBackgroundMode = NamePlateBackgroundMode.FixedColor;
-                    profile.NamePlateBackgroundR = 24;
-                    profile.NamePlateBackgroundG = 18;
-                    profile.NamePlateBackgroundB = 18;
-                    profile.NamePlateHealthBar = true;
-                    profile.NamePlateHealthBarOpacity = 100;
-                    profile.NamePlateOpacity = 82;
-                    profile.NamePlateBorderOpacity = 90;
-                    profile.NamePlateAvoidOverlap = true;
-                    profile.NamePlateHideAtFullHealth = false;
-                    profile.NamePlateHideAtFullHealthInWarmode = false;
-                    profile.NamePlateShowWordOfDeathIcon = false;
-                    profile.NamePlateFont = "avadonian";
-                    profile.NamePlateFontSize = 17;
-                    break;
-
-                case NamePlatePreset.WorldOfWarcraftRedName:
-                    profile.NamePlateUseFixedWidth = true;
-                    profile.NamePlateFixedWidth = 220;
-                    profile.NamePlateUseFixedHealthBarWidth = false;
-                    profile.NamePlateHealthBarFixedWidth = 220;
-                    profile.NamePlateHeight = 38;
-                    profile.NamePlateSplitHealthBar = true;
-                    profile.NamePlateCornerRadius = 1;
-                    profile.NamePlateHealthBarMode = NamePlateHealthBarMode.Green;
-                    profile.NamePlateBackgroundMode = NamePlateBackgroundMode.FixedColor;
-                    profile.NamePlateBackgroundR = 28;
-                    profile.NamePlateBackgroundG = 12;
-                    profile.NamePlateBackgroundB = 12;
-                    profile.NamePlateHealthBar = true;
-                    profile.NamePlateHealthBarOpacity = 100;
-                    profile.NamePlateOpacity = 85;
-                    profile.NamePlateBorderOpacity = 90;
-                    profile.NamePlateAvoidOverlap = true;
-                    profile.NamePlateHideAtFullHealth = false;
-                    profile.NamePlateHideAtFullHealthInWarmode = false;
-                    profile.NamePlateShowWordOfDeathIcon = false;
-                    profile.NamePlateFont = "avadonian";
-                    profile.NamePlateFontSize = 18;
-                    break;
-
-                default:
-                    break;
-            }
-
-            NameOverheadGump.InvalidateAllLayouts();
+            string page = GetPageString();
+            var refreshed = new ModernOptionsGump(World) { X = X, Y = Y };
+            refreshed.GoToPage(page);
+            Dispose();
+            UIManager.Add(refreshed);
         }
 
         private void BuildTazUO()
@@ -3595,6 +3461,7 @@ namespace ClassicUO.Game.UI.Gumps
 
             #region Nameplates
 
+            IReadOnlyList<NamePlatePresets.Entry> namePlatePresets = NamePlatePresets.GetEntries();
             page = ((int)PAGE.TUOOptions + 1003);
             content.AddToLeft(SubCategoryButton(TazLang.Get("nameplate_title", "Nameplates"), page, content.LeftWidth));
             content.ResetRightSide();
@@ -3613,12 +3480,21 @@ namespace ClassicUO.Game.UI.Gumps
                     TazLang.Get("nameplate_preset", "Preset"),
                     0,
                     ThemeSettings.COMBO_BOX_WIDTH,
-                    GetNamePlatePresetOptions(),
-                    (int)profile.NamePlatePreset,
-                    (i, s) => { ApplyNamePlatePreset((NamePlatePreset)i); },
+                    namePlatePresets.Select(p => p.Name).ToArray(),
+                    NamePlatePresets.GetSelectedIndex(profile, namePlatePresets),
+                    (i, s) => { ApplyNamePlatePreset(namePlatePresets[i]); },
                     false
                 ), true, page
             );
+
+            var saveNamePlatePreset = new ModernButton(0, 0, 200, 40, ButtonAction.Activate,
+                TazLang.Get("nameplate_savepreset"), ThemeSettings.BUTTON_FONT_COLOR);
+            saveNamePlatePreset.MouseUp += (_, e) =>
+            {
+                if (e.Button == MouseButtonType.Left)
+                    SaveNamePlatePresetWindow.Show(profile, RefreshNamePlateOptions);
+            };
+            content.AddToRight(saveNamePlatePreset, true, page);
 
             content.BlankLine();
 
@@ -3626,15 +3502,23 @@ namespace ClassicUO.Game.UI.Gumps
             (new CheckboxWithLabel(TazLang.Get("nameplate_healthbars", "Nameplates also act as health bars"), 0, profile.NamePlateHealthBar, (b) => { profile.NamePlateHealthBar = b; SetNamePlatePresetCustom(); }),
                 true, page);
 
+            content.AddToRight(
+                new CheckboxWithLabel(TazLang.Get("nameplate_notorietytext"), 0, profile.NamePlateUseNotorietyText,
+                    b => { profile.NamePlateUseNotorietyText = b; SetNamePlatePresetCustom(); }), true, page);
+            content.AddToRight(
+                new CheckboxWithLabel(TazLang.Get("nameplate_showmissinghealth"), 0, profile.NamePlateShowMissingHealth,
+                    b => { profile.NamePlateShowMissingHealth = b; SetNamePlatePresetCustom(); }), true, page);
+
             content.Indent();
 
             content.AddToRight
             (
-                new SliderWithLabel
+                c = new SliderWithLabel
                 (TazLang.Get("nameplate_hpopacity", "HP opacity"), 0, ThemeSettings.SLIDER_WIDTH, 0, 100, profile.NamePlateHealthBarOpacity,
                     (i) => { profile.NamePlateHealthBarOpacity = (byte)i; SetNamePlatePresetCustom(); }),
                 true, page
             );
+            c.SetTooltip(TazLang.Get("nameplate_resourceopacity_tooltip"));
 
             content.AddToRight
             (
@@ -3728,7 +3612,7 @@ namespace ClassicUO.Game.UI.Gumps
 
             content.AddToRight
             (
-                new ComboBoxWithLabel
+                c = new ComboBoxWithLabel
                 (
                     World,
                     TazLang.Get("nameplate_backgroundmode", "Background mode"),
@@ -3740,21 +3624,24 @@ namespace ClassicUO.Game.UI.Gumps
                     false
                 ), true, page
             );
+            c.SetTooltip(TazLang.Get("nameplate_backgroundmode_tooltip"));
 
             content.AddToRight
             (
-                new SliderWithLabel
+                c = new SliderWithLabel
                 (TazLang.Get("nameplate_borderopacity", "Border opacity"), 0, ThemeSettings.SLIDER_WIDTH, 0, 100, profile.NamePlateBorderOpacity,
                     (i) => { profile.NamePlateBorderOpacity = (byte)i; SetNamePlatePresetCustom(); }),
                 true, page
             );
+            c.SetTooltip(TazLang.Get("nameplate_borderopacity_tooltip"));
 
             content.AddToRight
             (
-                new SliderWithLabel
+                c = new SliderWithLabel
                 (TazLang.Get("nameplate_backgroundopacity", "Background opacity"), 0, ThemeSettings.SLIDER_WIDTH, 0, 100, profile.NamePlateOpacity,
                     (i) => { profile.NamePlateOpacity = (byte)i; SetNamePlatePresetCustom(); }), true, page
             );
+            c.SetTooltip(TazLang.Get("nameplate_backgroundopacity_tooltip"));
 
             content.AddToRight
             (